requirements.txt
python-pptx==0.6.21
fpdf2==2.7.6
streamlit==1.50.0
pandas==2.3.3
